refactor: 별을 별개 파일로 이동
